Cleanup App.cs
