+++ src/RepoCleaner/ConsoleComponents/OverviewTableRow.cs
@@ -2,7 +2,7 @@
 using Develix.RepoCleaner.Git.Model;
 using Spectre.Console;
 
-namespace Develix.RepoCleaner.ConsoleRenderer;
+namespace Develix.RepoCleaner.ConsoleComponents;
 
 internal class OverviewTableRow : OverviewTableRowBase
 {
