+++ src/RepoCleaner/ConsoleComponents/OverviewTableRowBase.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using Spectre.Console;
 
-namespace Develix.RepoCleaner.ConsoleRenderer;
+namespace Develix.RepoCleaner.ConsoleComponents;
 
 public abstract class OverviewTableRowBase
 {
