+++ src/RepoCleaner/ConsoleComponents/OverviewTableRowWithProject.cs
@@ -2,7 +2,7 @@
 using Develix.RepoCleaner.Git.Model;
 using Spectre.Console;
 
-namespace Develix.RepoCleaner.ConsoleRenderer;
+namespace Develix.RepoCleaner.ConsoleComponents;
 
 internal class OverviewTableRowWithProject : OverviewTableRow
 {
